% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Estimates the dependency network using association rules extracted from
x.}
\usage{
fit(x)
}
\arguments{
\item{x}{matrix input data used to fit the network}
}
\value{
fitted network
}
\description{
Estimates the dependency network using association rules extracted from
x.
}
\examples{
fit(iris[1:4])
}
